+++ src/rngs.h
@@ -1,16 +1,16 @@
 /* $Id: rngs.h 55 2005-09-13 22:29:52Z asminer $ */
-/* ----------------------------------------------------------------------- 
- * Name            : rngs.h  (header file for the library file rngs.c) 
+/* -----------------------------------------------------------------------
+ * Name            : rngs.h  (header file for the library file rngs.c)
  * Author          : Steve Park & Dave Geyer
  * Language        : ANSI C
  * Latest Revision : 09-22-98
- * ----------------------------------------------------------------------- 
+ * -----------------------------------------------------------------------
  */
 
 #if !defined( _RNGS_ )
 #define _RNGS_
 
-double Random(void);
+long double Random(void);
 void   PlantSeeds(long x);
 void   GetSeed(long *x);
 void   PutSeed(long x);
