+++ src/rvgs.c
@@ -1,6 +1,6 @@
 /* $Id: rvgs.c 55 2005-09-13 22:29:52Z asminer $ */
-/* -------------------------------------------------------------------------- 
- * This is an ANSI C library for generating random variates from six discrete 
+/* --------------------------------------------------------------------------
+ * This is an ANSI C library for generating random variates from six discrete
  * distributions
  *
  *      Generator         Range (x)     Mean         Variance
@@ -11,15 +11,15 @@
  *      Geometric(p)      x = 0,...     p/(1-p)      p/((1-p)*(1-p))
  *      Pascal(n, p)      x = 0,...     n*p/(1-p)    n*p/((1-p)*(1-p))
  *      Poisson(m)        x = 0,...     m            m
- * 
+ *
  * and seven continuous distributions
  *
- *      Uniform(a, b)     a < x < b     (a + b)/2    (b - a)*(b - a)/12 
+ *      Uniform(a, b)     a < x < b     (a + b)/2    (b - a)*(b - a)/12
  *      Exponential(m)    x > 0         m            m*m
  *      Erlang(n, b)      x > 0         n*b          n*b*b
  *      Normal(m, s)      all x         m            s*s
  *      Lognormal(a, b)   x > 0            see below
- *      Chisquare(n)      x > 0         n            2*n 
+ *      Chisquare(n)      x > 0         n            2*n
  *      Student(n)        all x         0  (n > 1)   n/(n - 2)   (n > 2)
  *
  * For the a Lognormal(a, b) random variable, the mean and variance are
@@ -39,23 +39,23 @@
 #include "rvgs.h"
 
 
-   long Bernoulli(double p)
+   long Bernoulli(long double p)
 /* ========================================================
- * Returns 1 with probability p or 0 with probability 1 - p. 
- * NOTE: use 0.0 < p < 1.0                                   
+ * Returns 1 with probability p or 0 with probability 1 - p.
+ * NOTE: use 0.0 < p < 1.0
  * ========================================================
- */ 
+ */
 {
   return ((Random() < (1.0 - p)) ? 0 : 1);
 }
 
-   long Binomial(long n, double p)
-/* ================================================================ 
- * Returns a binomial distributed integer between 0 and n inclusive. 
+   long Binomial(long n, long double p)
+/* ================================================================
+ * Returns a binomial distributed integer between 0 and n inclusive.
  * NOTE: use n > 0 and 0.0 < p < 1.0
  * ================================================================
  */
-{ 
+{
   long i, x = 0;
 
   for (i = 0; i < n; i++)
@@ -65,7 +65,7 @@
 
    long Equilikely(long a, long b)
 /* ===================================================================
- * Returns an equilikely distributed integer between a and b inclusive. 
+ * Returns an equilikely distributed integer between a and b inclusive.
  * NOTE: use a < b
  * ===================================================================
  */
@@ -73,7 +73,7 @@
   return (a + (long) ((b - a + 1) * Random()));
 }
 
-   long Geometric(double p)
+   long Geometric(long double p)
 /* ====================================================
  * Returns a geometric distributed non-negative integer.
  * NOTE: use 0.0 < p < 1.0
@@ -83,13 +83,13 @@
   return ((long) (log(1.0 - Random()) / log(p)));
 }
 
-   long Pascal(long n, double p)
-/* ================================================= 
- * Returns a Pascal distributed non-negative integer. 
+   long Pascal(long n, long double p)
+/* =================================================
+ * Returns a Pascal distributed non-negative integer.
  * NOTE: use n > 0 and 0.0 < p < 1.0
  * =================================================
  */
-{ 
+{
   long i, x = 0;
 
   for (i = 0; i < n; i++)
@@ -97,14 +97,14 @@
   return (x);
 }
 
-   long Poisson(double m)
-/* ================================================== 
- * Returns a Poisson distributed non-negative integer. 
+   long Poisson(long double m)
+/* ==================================================
+ * Returns a Poisson distributed non-negative integer.
  * NOTE: use m > 0
  * ==================================================
  */
-{ 
-  double t = 0.0;
+{
+  long double t = 0.0;
   long   x = 0;
 
   while (t < m) {
@@ -114,19 +114,19 @@
   return (x - 1);
 }
 
-   double Uniform(double a, double b)
-/* =========================================================== 
- * Returns a uniformly distributed real number between a and b. 
+   long double Uniform(long double a, long double b)
+/* ===========================================================
+ * Returns a uniformly distributed real number between a and b.
  * NOTE: use a < b
  * ===========================================================
  */
-{ 
+{
   return (a + (b - a) * Random());
 }
 
-   double Exponential(double m)
+   long double Exponential(long double m)
 /* =========================================================
- * Returns an exponentially distributed positive real number. 
+ * Returns an exponentially distributed positive real number.
  * NOTE: use m > 0.0
  * =========================================================
  */
@@ -134,37 +134,37 @@
   return (-m * log(1.0 - Random()));
 }
 
-   double Erlang(long n, double b)
-/* ================================================== 
+   long double Erlang(long n, long double b)
+/* ==================================================
  * Returns an Erlang distributed positive real number.
  * NOTE: use n > 0 and b > 0.0
  * ==================================================
  */
-{ 
+{
   long   i;
-  double x = 0.0;
+  long double x = 0.0;
 
-  for (i = 0; i < n; i++) 
+  for (i = 0; i < n; i++)
     x += Exponential(b);
   return (x);
 }
 
-   double Normal(double m, double s)
+   long double Normal(long double m, long double s)
 /* ========================================================================
  * Returns a normal (Gaussian) distributed real number.
  * NOTE: use s > 0.0
  *
- * Uses a very accurate approximation of the normal idf due to Odeh & Evans, 
+ * Uses a very accurate approximation of the normal idf due to Odeh & Evans,
  * J. Applied Statistics, 1974, vol 23, pp 96-97.
  * ========================================================================
  */
-{ 
-  const double p0 = 0.322232431088;     const double q0 = 0.099348462606;
-  const double p1 = 1.0;                const double q1 = 0.588581570495;
-  const double p2 = 0.342242088547;     const double q2 = 0.531103462366;
-  const double p3 = 0.204231210245e-1;  const double q3 = 0.103537752850;
-  const double p4 = 0.453642210148e-4;  const double q4 = 0.385607006340e-2;
-  double u, t, p, q, z;
+{
+  const long double p0 = 0.322232431088;     const long double q0 = 0.099348462606;
+  const long double p1 = 1.0;                const long double q1 = 0.588581570495;
+  const long double p2 = 0.342242088547;     const long double q2 = 0.531103462366;
+  const long double p3 = 0.204231210245e-1;  const long double q3 = 0.103537752850;
+  const long double p4 = 0.453642210148e-4;  const long double q4 = 0.385607006340e-2;
+  long double u, t, p, q, z;
 
   u   = Random();
   if (u < 0.5)
@@ -180,9 +180,9 @@
   return (m + s * z);
 }
 
-   double Lognormal(double a, double b)
-/* ==================================================== 
- * Returns a lognormal distributed positive real number. 
+   long double Lognormal(long double a, long double b)
+/* ====================================================
+ * Returns a lognormal distributed positive real number.
  * NOTE: use b > 0.0
  * ====================================================
  */
@@ -190,15 +190,15 @@
   return (exp(a + b * Normal(0.0, 1.0)));
 }
 
-   double Chisquare(long n)
+   long double Chisquare(long n)
 /* =====================================================
- * Returns a chi-square distributed positive real number. 
+ * Returns a chi-square distributed positive real number.
  * NOTE: use n > 0
  * =====================================================
  */
-{ 
+{
   long   i;
-  double z, x = 0.0;
+  long double z, x = 0.0;
 
   for (i = 0; i < n; i++) {
     z  = Normal(0.0, 1.0);
@@ -207,8 +207,8 @@
   return (x);
 }
 
-   double Student(long n)
-/* =========================================== 
+   long double Student(long n)
+/* ===========================================
  * Returns a student-t distributed real number.
  * NOTE: use n > 0
  * ===========================================
