+++ src/rvgs.h
@@ -1,29 +1,29 @@
 /* $Id: rvgs.h 55 2005-09-13 22:29:52Z asminer $ */
-/* ------------------------------------------------------------- 
+/* -------------------------------------------------------------
  * Name            : rvgs.h (header file for the library rvgs.c)
  * Author          : Steve Park & Dave Geyer
  * Language        : ANSI C
  * Latest Revision : 11-03-96
- * -------------------------------------------------------------- 
+ * --------------------------------------------------------------
  */
 
 #if !defined( _RVGS_ )
 #define _RVGS_
 
-long Bernoulli(double p);
-long Binomial(long n, double p);
+long Bernoulli(long double p);
+long Binomial(long n, long double p);
 long Equilikely(long a, long b);
-long Geometric(double p);
-long Pascal(long n, double p);
-long Poisson(double m);
+long Geometric(long double p);
+long Pascal(long n, long double p);
+long Poisson(long double m);
 
-double Uniform(double a, double b);
-double Exponential(double m);
-double Erlang(long n, double b);
-double Normal(double m, double s);
-double Lognormal(double a, double b);
-double Chisquare(long n);
-double Student(long n);
+long double Uniform(long double a, long double b);
+long double Exponential(long double m);
+long double Erlang(long n, long double b);
+long double Normal(long double m, long double s);
+long double Lognormal(long double a, long double b);
+long double Chisquare(long n);
+long double Student(long n);
 
 #endif
